restructure, use remoc to communicate between primary & secondaries
