+++ .github/workflows/rust.yml
@@ -34,4 +34,4 @@ jobs:
     - name: Run standalone tests
       run: cargo test -- --test-threads=1
     - name: Run cluster tests
-      run: cargo test --all-features --verbose
+      run: cargo test --all-features -- --test-threads=1
