+++ cluster-executor/src/request_providers.rs
@@ -1,166 +1,12 @@
-#![allow(clippy::upper_case_acronyms)]
-
-use crate::datagen::{generate_data, DataSchema};
-use crate::http_util::request::RequestSpecEnum;
-use crate::HttpReq;
 use anyhow::Result as AnyResult;
-use async_trait::async_trait;
-use futures_util::future::BoxFuture;
-use futures_util::stream::Stream;
-use http::Method;
-use log::{debug, trace, warn};
-use regex::Regex;
-use response_assert::ResponseAssertion;
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use smol_str::SmolStr;
+use datagen::generate_data;
+use log::{debug, trace};
+use overload_http::{HttpReq, RandomDataRequest, RequestFile, RequestList, RequestSpecEnum};
+use remoc::rtc::async_trait;
 use sqlx::sqlite::SqliteConnectOptions;
-use sqlx::{ConnectOptions, SqliteConnection};
-use std::cmp::{min, Ordering};
-use std::collections::{BinaryHeap, HashMap};
-use std::convert::TryFrom;
+use sqlx::ConnectOptions;
 use std::iter::repeat_with;
-use std::pin::Pin;
 use std::str::FromStr;
-use std::task::{Context, Poll};
-use std::time::Duration;
-use tokio_stream::StreamExt;
-
-const MAX_REQ_RET_SIZE: usize = 10;
-
-type ReqProvider = Box<dyn RequestProvider + Send>;
-
-pub enum ProviderOrFuture {
-    Provider(ReqProvider),
-    Future(BoxFuture<'static, (ReqProvider, AnyResult<Vec<HttpReq>>)>),
-    Dummy,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub enum Scheme {
-    HTTP,
-    //Unsupported
-    // HTTPS
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Target {
-    pub(crate) host: String,
-    pub(crate) port: u16,
-    pub(crate) protocol: Scheme,
-}
-
-/// Based on the QPS policy, generate requests to be sent to the application that's being tested.
-#[must_use = "futures do nothing unless polled"]
-#[allow(dead_code)]
-pub struct RequestGenerator {
-    // for how long test should run
-    duration: u32,
-    // How time should be scaled. Default value is 1, means will generate qps+request once
-    // every second.
-    // Doesn't support custom value for now.
-    time_scale: u8,
-    // total number of requests to be generated
-    total: u32,
-    // requests generated so far, initially 0
-    current_count: u32,
-    current_qps: u32,
-    shared_provider: bool,
-    pub(crate) target: Target,
-    // replace ProviderOrFuture::Provider with enum, no need to use conversion
-    // or check if Any can be used like sqlx AnyConnection
-    req_provider: String,
-    provider_or_future: ProviderOrFuture,
-    qps_scheme: Box<dyn RateScheme + Send>,
-    concurrent_connection: Box<dyn RateScheme + Send>,
-    pub(crate) response_assertion: Option<ResponseAssertion>,
-}
-
-impl RequestGenerator {
-    pub fn time_scale(&self) -> u8 {
-        self.time_scale
-    }
-
-    pub fn shared_provider(&self) -> bool {
-        self.shared_provider
-    }
-
-    pub fn request_provider(&self) -> &String {
-        &self.req_provider
-    }
-}
-
-pub trait RateScheme {
-    fn next(&mut self, nth: u32, last_qps: Option<u32>) -> u32;
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ConstantRate {
-    pub count_per_sec: u32,
-}
-
-impl RateScheme for ConstantRate {
-    #[inline]
-    fn next(&mut self, _nth: u32, _last_qps: Option<u32>) -> u32 {
-        self.count_per_sec
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ArraySpec {
-    count_per_sec: Vec<u32>,
-}
-
-impl ArraySpec {
-    pub fn new(count_per_sec: Vec<u32>) -> Self {
-        Self { count_per_sec }
-    }
-}
-
-impl RateScheme for ArraySpec {
-    #[inline]
-    fn next(&mut self, nth: u32, _last_qps: Option<u32>) -> u32 {
-        let len = self.count_per_sec.len();
-        if len != 0 {
-            let idx = nth as usize % len;
-            let val = self.count_per_sec.get(idx).unwrap();
-            *val
-        } else {
-            0
-        }
-    }
-}
-
-/// Used only for concurrent connections, creates a pool of maximum size specified
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Bounded {
-    max: u32,
-}
-
-impl Default for Bounded {
-    fn default() -> Self {
-        Bounded { max: 500 }
-    }
-}
-
-impl RateScheme for Bounded {
-    fn next(&mut self, _nth: u32, _last_value: Option<u32>) -> u32 {
-        self.max
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-pub struct Elastic {
-    max: u32,
-}
-
-impl RateScheme for Elastic {
-    fn next(&mut self, _nth: u32, _last_qps: Option<u32>) -> u32 {
-        0
-    }
-}
 
 #[async_trait]
 pub trait RequestProvider {
@@ -173,6 +19,10 @@ pub trait RequestProvider {
     /// should be handled properly
     async fn get_n(&mut self, n: usize) -> AnyResult<Vec<HttpReq>>;
 
+    // /// Generate `n` requests and put into provided vec.
+    // /// Returns number of requests generated successfully
+    // async fn fill_vec(&mut self, vec: &mut Vec<HttpReq>, n: usize) -> AnyResult<usize>;
+
     /// number of total requests
     fn size_hint(&self) -> usize;
 
@@ -181,12 +31,8 @@ pub trait RequestProvider {
     fn shared(&self) -> bool;
 
     /// Not a good solution; it creates circular dependency, temporary hack, should find better solution
+    #[deprecated]
     fn to_json_str(&self) -> String;
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct RequestList {
-    pub(crate) data: Vec<HttpReq>,
 }
 
 #[async_trait]
@@ -222,39 +68,6 @@ impl RequestProvider for RequestList {
             data: self.data.clone(),
         });
         serde_json::to_string(&spec_enum).unwrap()
-    }
-}
-
-impl From<Vec<HttpReq>> for RequestList {
-    fn from(data: Vec<HttpReq>) -> Self {
-        RequestList { data }
-    }
-}
-
-/// Test request with file data
-#[derive(Debug, Serialize, Deserialize)]
-pub struct RequestFile {
-    pub(crate) file_name: String,
-    #[serde(skip)]
-    inner: Option<SqliteConnection>,
-    #[serde(skip)]
-    #[serde(default = "default_request_file_size")]
-    size: usize,
-}
-
-impl RequestFile {
-    pub(crate) fn new(file_name: String) -> RequestFile {
-        RequestFile {
-            file_name,
-            inner: None,
-            size: default_request_file_size(),
-        }
-    }
-}
-
-impl Clone for RequestFile {
-    fn clone(&self) -> Self {
-        RequestFile::new(self.file_name.clone())
     }
 }
 
@@ -317,51 +130,6 @@ impl RequestProvider for RequestFile {
         serde_json::to_string(&spec_enum).unwrap()
     }
 }
-#[derive(Debug, Serialize, Deserialize, Clone)]
-struct UrlParam {
-    name: SmolStr,
-    start: usize,
-    end: usize,
-}
-
-impl Eq for UrlParam {}
-
-impl PartialEq<Self> for UrlParam {
-    fn eq(&self, other: &Self) -> bool {
-        self.start.eq(&other.start)
-    }
-}
-
-impl PartialOrd<Self> for UrlParam {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.start.partial_cmp(&other.start)
-    }
-}
-
-impl Ord for UrlParam {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.start.cmp(&other.start)
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-#[allow(clippy::type_complexity)]
-pub struct RandomDataRequest {
-    #[serde(skip)]
-    #[serde(default)]
-    init: bool,
-    #[serde(with = "http_serde::method")]
-    pub method: Method,
-    //todo as a http::Uri
-    pub url: String,
-    #[serde(skip)]
-    url_param_pos: Option<BinaryHeap<UrlParam>>,
-    #[serde(default = "HashMap::new")]
-    pub headers: HashMap<String, String>,
-    body_schema: Option<DataSchema>,
-    uri_param_schema: Option<DataSchema>,
-}
 
 #[async_trait]
 impl RequestProvider for RandomDataRequest {
@@ -384,7 +152,7 @@ impl RequestProvider for RandomDataRequest {
                     RandomDataRequest::substitute_param_with_data(&mut url, positions, data)
                 }
             }
-            if matches!(self.method, Method::POST) {
+            if matches!(self.method, http::Method::POST) {
                 if let Some(schema) = &self.body_schema {
                     body = Some(Vec::from(generate_data(schema).to_string().as_bytes()));
                 }
@@ -414,286 +182,14 @@ impl RequestProvider for RandomDataRequest {
     }
 }
 
-impl RandomDataRequest {
-    fn find_param_positions(url: &str) -> BinaryHeap<UrlParam> {
-        let re = Regex::new("\\{[a-zA-Z0-9_-]+}").unwrap();
-        // let url_str = url;
-        let matches = re.find_iter(url);
-        let mut match_indices = BinaryHeap::new();
-        for m in matches {
-            match_indices.push(UrlParam {
-                name: SmolStr::new(&url[m.start() + 1..m.end() - 1]),
-                start: m.start(),
-                end: m.end(),
-            });
-        }
-        match_indices
-    }
-
-    fn substitute_param_with_data(url: &mut String, positions: &BinaryHeap<UrlParam>, data: Value) {
-        for url_param in positions.iter() {
-            if let Some(p_val) = data.get(&url_param.name.as_str()) {
-                let val: String = if p_val.is_string() {
-                    String::from(p_val.as_str().unwrap())
-                } else {
-                    p_val
-                        .as_i64()
-                        .map_or(String::from("unknown"), |t| t.to_string())
-                };
-                url.replace_range(url_param.start..url_param.end as usize, &val);
-            }
-        }
-    }
-}
-
-fn default_request_file_size() -> usize {
-    999
-}
-
-//todo use failure rate for adaptive qps control
-//https://micrometer.io/docs/concepts#rate-aggregation
-
-impl RequestGenerator {
-    pub fn new(
-        duration: u32,
-        requests: ReqProvider,
-        qps_scheme: Box<dyn RateScheme + Send>,
-        target: Target,
-        concurrent_connection: Option<Box<dyn RateScheme + Send>>,
-        response_assertion: Option<ResponseAssertion>,
-    ) -> Self {
-        let time_scale: u8 = 1;
-        let total = duration * time_scale as u32;
-        let concurrent_connection =
-            concurrent_connection.unwrap_or_else(|| Box::new(Elastic::default()));
-        RequestGenerator {
-            duration,
-            time_scale,
-            total,
-            current_qps: 0,
-            shared_provider: requests.shared(),
-            req_provider: requests.to_json_str(),
-            provider_or_future: ProviderOrFuture::Provider(requests),
-            qps_scheme,
-            current_count: 0,
-            target,
-            concurrent_connection,
-            response_assertion,
-        }
-    }
-}
-
-impl Stream for RequestGenerator {
-    type Item = (u32, AnyResult<Vec<HttpReq>>, u32);
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        if self.current_count >= self.total {
-            trace!("finished generating QPS");
-            Poll::Ready(None)
-        } else {
-            let provider_or_future =
-                std::mem::replace(&mut self.provider_or_future, ProviderOrFuture::Dummy);
-            match provider_or_future {
-                ProviderOrFuture::Provider(mut provider) => {
-                    let current_count = self.current_count;
-                    let qps = self.qps_scheme.next(current_count, None);
-                    let len = provider.size_hint();
-                    let req_size = min(qps as usize, min(len, MAX_REQ_RET_SIZE));
-                    trace!("request size: {}, qps: {}, len: {}", req_size, qps, len);
-                    let provider_and_fut = async move {
-                        let requests = if req_size > 0 {
-                            provider.get_n(req_size).await
-                        } else {
-                            Ok(vec![])
-                        };
-                        (provider, requests)
-                    };
-                    self.provider_or_future = ProviderOrFuture::Future(Box::pin(provider_and_fut));
-                    self.current_qps = qps;
-                    self.poll_next(cx)
-                }
-                ProviderOrFuture::Future(mut future) => match future.as_mut().poll(cx) {
-                    Poll::Ready((provider, result)) => {
-                        let qps = self.current_qps;
-                        self.current_qps = 0;
-                        let current_count = self.current_count;
-                        let conn = self.concurrent_connection.next(current_count, None);
-                        self.current_count += 1;
-                        self.provider_or_future = ProviderOrFuture::Provider(provider);
-                        Poll::Ready(Some((qps, result, conn)))
-                    }
-                    Poll::Pending => {
-                        self.provider_or_future = ProviderOrFuture::Future(future);
-                        Poll::Pending
-                    }
-                },
-                ProviderOrFuture::Dummy => panic!("Something went wrong :("),
-            }
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (0, Some(self.total as usize))
-    }
-}
-
-#[allow(clippy::let_and_return)]
-pub fn request_generator_stream(
-    generator: RequestGenerator,
-) -> impl Stream<Item = (u32, AnyResult<Vec<HttpReq>>, u32)> {
-    //todo impl throttled stream function according to time_scale
-    let scale = generator.time_scale;
-    let throttle = generator.throttle(Duration::from_millis(1000 / scale as u64));
-    // tokio::pin!(throttle); //Why causes error => the parameter type `Q` may not live long enough.
-    throttle
-}
-
-/// Increase QPS linearly as per equation
-/// `y=ceil(ax+b)` until hit the max cap
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Linear {
-    a: f32,
-    b: u32,
-    max: u32,
-}
-
-impl RateScheme for Linear {
-    fn next(&mut self, nth: u32, _last_qps: Option<u32>) -> u32 {
-        min((self.a * nth as f32).ceil() as u32 + self.b, self.max)
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(try_from = "StepShadowType")]
-pub struct Step {
-    pub(crate) start: u32,
-    pub(crate) end: u32,
-    pub(crate) rate: u32,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(try_from = "StepsShadowType")]
-pub struct Steps {
-    #[serde(skip)]
-    current_step: usize,
-    steps: Vec<Step>,
-}
-
-impl RateScheme for Steps {
-    fn next(&mut self, nth: u32, _last_qps: Option<u32>) -> u32 {
-        let x = if let Some(step) = self.steps.get(self.current_step) {
-            if step.start <= nth && step.end >= nth {
-                (step.rate, false)
-            } else {
-                //not found in current_step, check next
-                self.steps.get(self.current_step + 1).map_or_else(
-                    || {
-                        (
-                            //not found in current_step+1, use last step as fallback
-                            self.steps.last().map_or(0, |s| s.rate),
-                            // (self.steps.len() - 1) - self.current_step,
-                            false,
-                        )
-                    },
-                    |s| (s.rate, true),
-                )
-            }
-        } else {
-            warn!("Invalid current step");
-            (0, false)
-        };
-        if x.1 {
-            self.current_step += 1;
-        }
-        x.0
-    }
-}
-
-#[derive(Deserialize)]
-pub struct StepShadowType {
-    start: u32,
-    end: u32,
-    rate: u32,
-}
-
-impl TryFrom<StepShadowType> for Step {
-    type Error = anyhow::Error;
-
-    fn try_from(value: StepShadowType) -> Result<Self, Self::Error> {
-        if value.start >= value.end {
-            return Err(anyhow::anyhow!(
-                "start({}) can not be greater than or equal to end({})",
-                value.start,
-                value.end
-            ));
-        }
-        Ok(Step {
-            start: value.start,
-            end: value.end,
-            rate: value.rate,
-        })
-    }
-}
-
-#[derive(Deserialize)]
-pub struct StepsShadowType {
-    steps: Vec<Step>,
-}
-
-impl TryFrom<StepsShadowType> for Steps {
-    type Error = anyhow::Error;
-
-    fn try_from(value: StepsShadowType) -> Result<Self, Self::Error> {
-        let mut value = value;
-        if value.steps.is_empty() {
-            return Err(anyhow::anyhow!("No steps found."));
-        }
-        if value.steps.len() == 1 {
-            return Err(anyhow::anyhow!(
-                "Need more than one step. Use ConstantQPS for single step."
-            ));
-        }
-        value.steps.sort_by(|a, b| a.start.cmp(&b.start));
-        //check continuity of steps
-        let steps = &value.steps;
-        let mut iter = steps.iter();
-        let first = iter.next().unwrap();
-        if first.start != 0 {
-            return Err(anyhow::anyhow!(
-                "Steps should start from 0, but starts from {} instead.",
-                first.start
-            ));
-        }
-
-        let mut prev = first.end;
-        for current in iter {
-            if prev + 1 != current.start {
-                return Err(anyhow::anyhow!("Steps are not continuous. An step ends at {} but next step should start at {}, but starts at {}",
-                 prev, prev+1, current.start));
-            } else {
-                prev = current.end;
-            }
-        }
-        Ok(Steps {
-            current_step: 0,
-            steps: value.steps,
-        })
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod test {
-    use crate::datagen::{data_schema_from_value, generate_data};
-    use crate::generator::RateScheme;
-    use crate::generator::Scheme;
-    use crate::generator::Target;
-    use crate::generator::{
-        request_generator_stream, ConstantRate, RandomDataRequest, RequestFile, RequestGenerator,
-        RequestList, RequestProvider, MAX_REQ_RET_SIZE,
-    };
-    use crate::generator::{Linear, Steps};
-    use crate::HttpReq;
+    use super::*;
+    use crate::rate_spec::RateScheme;
+    use crate::RequestGenerator;
+    use datagen::data_schema_from_value;
     use http::Method;
+    use overload_http::{ConstantRate, Linear, Scheme, Steps, Target};
     use regex::Regex;
     use serde_json::Value;
     use std::collections::HashMap;
@@ -702,9 +198,7 @@ pub(crate) mod test {
     use std::time::Duration;
     use tokio::fs::File;
     use tokio::io::AsyncWriteExt;
-    use tokio::time;
     use tokio_stream::StreamExt;
-    use tokio_test::{assert_pending, assert_ready, task};
     use uuid::Uuid;
 
     static ONCE: Once = Once::new();
@@ -834,7 +328,7 @@ pub(crate) mod test {
         let ret = generator.next().await;
         if let Some(arg) = ret {
             assert_eq!(arg.0, 3);
-            assert_eq!(arg.1.unwrap().len(), 0);
+            assert_eq!(arg.1, 0);
         } else {
             panic!("fails");
         }
@@ -860,12 +354,11 @@ pub(crate) mod test {
         );
         let ret = generator.next().await;
         if let Some(arg) = ret {
-            let requests = arg.1.unwrap();
             assert_eq!(arg.0, 3);
-            assert_eq!(requests.len(), 1);
-            if let Some(req) = requests.get(0) {
-                assert_eq!(req.method, http::Method::GET);
-            }
+            assert_eq!(arg.1, 0);
+            // if let Some(req) = requests.get(0) {
+            //     assert_eq!(req.method, http::Method::GET);
+            // }
         } else {
             panic!()
         }
@@ -888,7 +381,7 @@ pub(crate) mod test {
         let ret = generator.next().await;
         if let Some(arg) = ret {
             assert_eq!(arg.0, 3);
-            assert_eq!(arg.1.unwrap().len(), 3);
+            assert_eq!(arg.1, 0);
         } else {
             panic!()
         }
@@ -911,7 +404,7 @@ pub(crate) mod test {
         let ret = generator.next().await;
         if let Some(arg) = ret {
             assert_eq!(arg.0, 15);
-            assert_eq!(arg.1.unwrap().len(), MAX_REQ_RET_SIZE);
+            assert_eq!(arg.1, 0);
         } else {
             panic!()
         }
@@ -934,19 +427,19 @@ pub(crate) mod test {
         );
         let arg = generator.next().await.unwrap();
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 0);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 0);
         tokio::time::pause();
         let arg = generator.next().await.unwrap();
         tokio::time::advance(Duration::from_millis(1001)).await;
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 0);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 0);
         let arg = generator.next().await.unwrap();
         tokio::time::advance(Duration::from_millis(1001)).await;
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 0);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 0);
         let arg = generator.next().await;
         tokio::time::advance(Duration::from_millis(1001)).await;
         assert!(arg.is_none())
@@ -974,8 +467,8 @@ pub(crate) mod test {
         let arg = generator.next().await.unwrap();
         //assert first element
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 1);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 1);
         tokio::time::pause();
 
         let _arg = generator.next().await.unwrap();
@@ -985,8 +478,8 @@ pub(crate) mod test {
         tokio::time::advance(Duration::from_millis(1001)).await;
         // assert third element
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 2);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 2);
 
         for _ in 0..26 {
             let arg = generator.next().await;
@@ -996,37 +489,37 @@ pub(crate) mod test {
         let arg = generator.next().await.unwrap();
         tokio::time::advance(Duration::from_millis(1001)).await;
         assert_eq!(arg.0, 3);
-        assert_eq!(arg.1.unwrap().len(), 3);
-        assert_eq!(arg.2, 10);
+        // assert_eq!(arg.1.unwrap().len(), 3);
+        assert_eq!(arg.1, 10);
         let arg = generator.next().await;
         tokio::time::advance(Duration::from_millis(1001)).await;
         assert!(arg.is_none());
     }
 
     //noinspection Duplicates
-    #[tokio::test]
-    async fn test_request_generator_stream() {
-        time::pause();
-        let generator = RequestGenerator::new(
-            3,
-            Box::new(req_list_with_n_req(1)),
-            Box::new(ConstantRate { count_per_sec: 3 }),
-            Target {
-                host: "example.com".into(),
-                port: 8080,
-                protocol: Scheme::HTTP,
-            },
-            None,
-            None,
-        );
-        let throttle = request_generator_stream(generator);
-        let mut throttle = task::spawn(throttle);
-        let ret = throttle.poll_next();
-        assert_ready!(ret);
-        assert_pending!(throttle.poll_next());
-        time::advance(Duration::from_millis(1001)).await;
-        assert_ready!(throttle.poll_next());
-    }
+    // #[tokio::test]
+    // async fn test_request_generator_stream() {
+    //     time::pause();
+    //     let generator = RequestGenerator::new(
+    //         3,
+    //         Box::new(req_list_with_n_req(1)),
+    //         Box::new(ConstantRate { count_per_sec: 3 }),
+    //         Target {
+    //             host: "example.com".into(),
+    //             port: 8080,
+    //             protocol: Scheme::HTTP,
+    //         },
+    //         None,
+    //         None,
+    //     );
+    //     let throttle = request_generator_stream(generator);
+    //     let mut throttle = task::spawn(throttle);
+    //     let ret = throttle.poll_next();
+    //     assert_ready!(ret);
+    //     assert_pending!(throttle.poll_next());
+    //     time::advance(Duration::from_millis(1001)).await;
+    //     assert_ready!(throttle.poll_next());
+    // }
 
     pub fn req_list_with_n_req(n: usize) -> RequestList {
         let data = (0..n)
