+++ datagen/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "datagen"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+
+[dependencies.regex]
+version = "1.4"
+
+[dependencies.serde]
+version = "1.0"
+features = ["derive"]
+
+[dependencies.serde_json]
+version = "1.0"
+
+[dependencies.fastrand]
+version = "1.4"
+
+[dependencies.rand]
+version = "0.8"
+
+[dependencies.regex-syntax]
+version = "0.6"
+
+[dependencies.regex_generate]
+git = "https://github.com/eipi1/regex_generate"
+branch = "pub-gen-from-hir"
+
+[dependencies.anyhow]
+version = "1.0"
+
+
+[dependencies.log]
+version = "0.4"
+
+[dev-dependencies.assert-json-diff]
+version = "2.0"
+
