+++ overload-http/Cargo.toml
@@ -1,0 +1,67 @@
+[package]
+name = "overload-http"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies.response-assert]
+path = "../response-assert"
+
+[dependencies.datagen]
+path="../datagen"
+
+[dependencies.async-trait]
+version = "0.1"
+
+[dependencies.serde]
+version = "1.0"
+features = ["derive"]
+
+[dependencies.serde_json]
+version = "1.0"
+
+[dependencies.http-serde]
+version = "1.0"
+
+[dependencies.smallvec]
+version = "1.7"
+features = ["serde", "union", "const_generics"]
+
+[dependencies.anyhow]
+version = "1.0"
+
+[dependencies.log]
+version = "0.4"
+
+[dependencies.fastrand]
+version = "1.4"
+
+[dependencies.sqlx]
+version = "0.5"
+default-features = false
+features = ["sqlite", "runtime-tokio-native-tls"]
+
+[dependencies.once_cell]
+version = "1.16"
+
+[dependencies.smol_str]
+version = "0.1.18"
+features = ["serde"]
+
+[dependencies.http]
+version = "0.2"
+
+[dependencies.hyper]
+version = "0.14"
+
+[dependencies.regex]
+version = "1.4"
+
+[dependencies.uuid]
+version = "1.1"
+features = ["v4"]
+
+[dev-dependencies.tokio]
+version = "1.19"
+features = ["rt-multi-thread"]
