+++ overload/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "overload"
-version = "0.5.0"
+version = "0.6.0"
 authors = ["Md Sarowar Alam <sarowar.cse@gmail.com>"]
 edition = "2018"
 license = "MIT"
@@ -19,16 +19,28 @@ path = "src/overload/lib.rs"
 [dependencies.response-assert]
 path = "../response-assert"
 
+[dependencies.overload-metrics]
+path = "../overload-metrics"
+
+[dependencies.overload-http]
+path = "../overload-http"
+
+[dependencies.cluster-executor]
+path = "../cluster-executor"
+
+[dependencies.datagen]
+path = "../datagen"
+
 [dependencies.rust-cloud-discovery]
-version = "0.1"
+version = "0.2"
 optional = true
 
 [dependencies.cloud-discovery-kubernetes]
-version = "0.1"
+version = "0.2"
 optional = true
 
 [dependencies.cluster-mode]
-version = "0.1"
+version = "0.2"
 optional = true
 
 [dependencies.bytes]
@@ -187,8 +199,4 @@ version = "0.1"
 version = "2.2"
 
 [features]
-default = ["debug"]
-
 cluster = ["rust-cloud-discovery", "cloud-discovery-kubernetes", "cluster-mode"]
-# addtional checks for debugging/logging purpose
-debug = []
